# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Simix API - virtual phone numbers and mobile money wallet
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: services
    description: Available SMS services
  - name: countries
    description: Available countries
  - name: numbers
    description: Virtual phone numbers
  - name: wallet
    description: Wallet, recharges, transactions
  - name: dashboard
    description: Dashboard summaries

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "200":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with phone or username + password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout the current session
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get the currently authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all available SMS services (WhatsApp, Telegram, etc.)
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: category
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Service list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"

  /services/popular:
    get:
      operationId: listPopularServices
      tags: [services]
      summary: List popular services for the dashboard
      responses:
        "200":
          description: Popular services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"

  /countries:
    get:
      operationId: listCountries
      tags: [countries]
      summary: List available countries with availability and pricing
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: serviceId
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Country list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Country"

  /countries/popular:
    get:
      operationId: listPopularCountries
      tags: [countries]
      summary: List popular countries for the dashboard
      responses:
        "200":
          description: Popular countries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Country"

  /numbers/quote:
    get:
      operationId: getNumberQuote
      tags: [numbers]
      summary: Get pricing/availability quote for a service+country combo
      parameters:
        - name: serviceId
          in: query
          required: true
          schema:
            type: string
        - name: countryId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NumberQuote"

  /numbers:
    post:
      operationId: requestNumber
      tags: [numbers]
      summary: Buy/assign a virtual number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestNumberInput"
      responses:
        "200":
          description: Number assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VirtualNumber"

  /numbers/active:
    get:
      operationId: listActiveNumbers
      tags: [numbers]
      summary: List the user's currently active numbers
      responses:
        "200":
          description: Active numbers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VirtualNumber"

  /numbers/history:
    get:
      operationId: listNumberHistory
      tags: [numbers]
      summary: List the user's full number history
      responses:
        "200":
          description: History
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VirtualNumber"

  /numbers/{numberId}:
    get:
      operationId: getNumber
      tags: [numbers]
      summary: Get a virtual number by id (with messages)
      parameters:
        - name: numberId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Virtual number details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VirtualNumber"

  /numbers/{numberId}/messages:
    get:
      operationId: listNumberMessages
      tags: [numbers]
      summary: Poll messages received for a number
      parameters:
        - name: numberId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SmsMessage"

  /numbers/{numberId}/extend:
    post:
      operationId: extendNumber
      tags: [numbers]
      summary: Extend the active time of a virtual number by 5 minutes
      parameters:
        - name: numberId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Extended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VirtualNumber"

  /numbers/{numberId}/cancel:
    post:
      operationId: cancelNumber
      tags: [numbers]
      summary: Cancel a virtual number and refund if no SMS received
      parameters:
        - name: numberId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VirtualNumber"

  /wallet:
    get:
      operationId: getWallet
      tags: [wallet]
      summary: Get the user's wallet balance
      responses:
        "200":
          description: Wallet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Wallet"

  /wallet/recharge:
    post:
      operationId: rechargeWallet
      tags: [wallet]
      summary: Recharge the wallet with a chosen payment method
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RechargeInput"
      responses:
        "200":
          description: Recharge processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"

  /wallet/transactions:
    get:
      operationId: listTransactions
      tags: [wallet]
      summary: List the user's wallet transactions
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"

  /wallet/payment-methods:
    get:
      operationId: listPaymentMethods
      tags: [wallet]
      summary: List supported payment methods for recharges
      responses:
        "200":
          description: Payment methods
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentMethod"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Aggregated dashboard summary (balance, totals, recent activity)
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean

    User:
      type: object
      required: [id, fullName, phone, email, balance, verified, createdAt]
      properties:
        id:
          type: string
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        username:
          type: string
        balance:
          type: integer
          description: Balance in FCFA (whole units)
        verified:
          type: boolean
        status:
          type: string
          description: e.g. Standard / Premium
        createdAt:
          type: string
          format: date-time
        totalSpent:
          type: integer
        transactionsCount:
          type: integer

    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string

    RegisterInput:
      type: object
      required: [fullName, phone, password]
      properties:
        fullName:
          type: string
          minLength: 2
        phone:
          type: string
          minLength: 6
        countryCode:
          type: string
        password:
          type: string
          minLength: 6
        email:
          type: string

    LoginInput:
      type: object
      required: [identifier, password]
      properties:
        identifier:
          type: string
          description: phone number or username
        password:
          type: string
        method:
          type: string
          enum: [phone, username]

    Service:
      type: object
      required: [id, name, slug, scope, price, available, color]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
        scope:
          type: string
          description: e.g. Global
        price:
          type: integer
          description: Price in FCFA
        available:
          type: integer
          description: Number of available numbers
        color:
          type: string
          description: Hex color used for the icon background
        category:
          type: string
        popular:
          type: boolean

    Country:
      type: object
      required: [id, name, code, dialCode, flag, available, price]
      properties:
        id:
          type: string
        name:
          type: string
        code:
          type: string
          description: ISO country code
        dialCode:
          type: string
          description: e.g. +1, +33
        flag:
          type: string
          description: Emoji flag
        available:
          type: integer
        price:
          type: integer
          description: Price in FCFA
        popular:
          type: boolean

    NumberQuote:
      type: object
      required: [service, country, available, waitTime, price, fees, total, validityMinutes]
      properties:
        service:
          $ref: "#/components/schemas/Service"
        country:
          $ref: "#/components/schemas/Country"
        available:
          type: integer
        waitTime:
          type: string
          description: e.g. "20 - 60 sec"
        price:
          type: integer
        fees:
          type: integer
        total:
          type: integer
        validityMinutes:
          type: integer

    RequestNumberInput:
      type: object
      required: [serviceId, countryId]
      properties:
        serviceId:
          type: string
        countryId:
          type: string

    VirtualNumber:
      type: object
      required: [id, phoneNumber, status, expiresAt, createdAt, service, country, price, messages]
      properties:
        id:
          type: string
        phoneNumber:
          type: string
        status:
          type: string
          enum: [waiting, received, expired, cancelled]
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        price:
          type: integer
        service:
          $ref: "#/components/schemas/Service"
        country:
          $ref: "#/components/schemas/Country"
        messages:
          type: array
          items:
            $ref: "#/components/schemas/SmsMessage"

    SmsMessage:
      type: object
      required: [id, sender, body, code, receivedAt]
      properties:
        id:
          type: string
        sender:
          type: string
        body:
          type: string
        code:
          type: string
        receivedAt:
          type: string
          format: date-time

    Wallet:
      type: object
      required: [balance, currency]
      properties:
        balance:
          type: integer
        currency:
          type: string

    PaymentMethod:
      type: object
      required: [id, name, slug, description, color]
      properties:
        id:
          type: string
        name:
          type: string
        slug:
          type: string
          enum: [mobile_money, card, bank_transfer, orange_money, mtn_money]
        description:
          type: string
        color:
          type: string
        recommended:
          type: boolean

    RechargeInput:
      type: object
      required: [amount, methodSlug]
      properties:
        amount:
          type: integer
          minimum: 100
        methodSlug:
          type: string

    Transaction:
      type: object
      required: [id, type, amount, status, createdAt]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [recharge, purchase, refund]
        amount:
          type: integer
        status:
          type: string
          enum: [pending, completed, failed]
        method:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time

    DashboardSummary:
      type: object
      required: [balance, currency, totalNumbers, totalSpent, recentNumbers, recentTransactions]
      properties:
        balance:
          type: integer
        currency:
          type: string
        totalNumbers:
          type: integer
        totalSpent:
          type: integer
        recentNumbers:
          type: array
          items:
            $ref: "#/components/schemas/VirtualNumber"
        recentTransactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"