# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Simix API - virtual phone numbers and mobile money wallet
 * OpenAPI spec version: 0.1.0
 */
import type { Transaction } from "./transaction";
import type { VirtualNumber } from "./virtualNumber";

export interface DashboardSummary {
  balance: number;
  currency: string;
  totalNumbers: number;
  totalSpent: number;
  recentNumbers: VirtualNumber[];
  recentTransactions: Transaction[];
}