# lib/api-zod/src/generated/types/paymentMethod.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Simix API - virtual phone numbers and mobile money wallet
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentMethodSlug } from "./paymentMethodSlug";

export interface PaymentMethod {
  id: string;
  name: string;
  slug: PaymentMethodSlug;
  description: string;
  color: string;
  recommended?: boolean;
}