# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Simix API - virtual phone numbers and mobile money wallet
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: string;
  name: string;
  slug: string;
  /** e.g. Global */
  scope: string;
  /** Price in FCFA */
  price: number;
  /** Number of available numbers */
  available: number;
  /** Hex color used for the icon background */
  color: string;
  category?: string;
  popular?: boolean;
}