# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Simix API - virtual phone numbers and mobile money wallet
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: string;
  fullName: string;
  phone: string;
  email: string;
  username?: string;
  /** Balance in FCFA (whole units) */
  balance: number;
  verified: boolean;
  /** e.g. Standard / Premium */
  status?: string;
  createdAt: Date;
  totalSpent?: number;
  transactionsCount?: number;
}